'Move sony store methods outside of Client.js'
